Converted the reachable regions to circle unions
I changed the reachable regions to be represented as the intersection of two circles rather than as convex hulls. In addition, I updated the test scripts in order to reflect this change. One important note is that I added the AddLorentzConeConstraint binding to drake in order to add these constraints. This is necessary and is not currently in upstream drake.